Fixed looping problem with 409 from create job
When drpy would get a 409 trying to create a job it would get stuck in a loop instead of marking the currentjob
as failed and making a new job. This change fixes rackn/provision-server#189

This change also includes a basic Dockerfile that can be used to run drpy during development. Instructions can
be found in the scripts/README.md

Signed-off-by: Michael Rice <michael@michaelrice.org>